+++ apps/meteor/app/apps/server/bridges/livechat.ts
@@ -14,6 +14,7 @@ import { Livechat as LivechatTyped } from '../../../livechat/server/lib/Livechat
 import { closeRoom } from '../../../livechat/server/lib/closeRoom';
 import { getRoomMessages } from '../../../livechat/server/lib/getRoomMessages';
 import type { ILivechatMessage } from '../../../livechat/server/lib/localTypes';
+import { updateMessage } from '../../../livechat/server/lib/messages';
 import { settings } from '../../../settings/server';
 
 declare module '@rocket.chat/apps/dist/converters/IAppMessagesConverter' {
@@ -89,7 +90,7 @@ export class AppLivechatBridge extends LivechatBridge {
 		};
 
 		// @ts-expect-error IVisitor vs ILivechatVisitor :(
-		await LivechatTyped.updateMessage(data);
+		await updateMessage(data);
 	}
 
 	protected async createRoom(
